Eliminate msg unpacking on proxy
The message unpacking on the proxy
has proven to be unnecessary. This is
a fairly slow operation, so we should
avoid doing it.

Additionally, packing and unpacking
the message has security implications
once we move to trusted messaging.

Also removes the passing of msg_id by
various methods as it was redundant
(msg_id, when needed was part of the msg)

Change-Id: Idbc9291fcce6ff6f6108573811d065a5c2c24255