In-process ZmqProxy option.
Using impl_zmq requires running once instance of
ZmqProxy (via bin/oslo-zmq-receiver) per host.

When an operator knows that only a single service
will be operating on a host (nova-compute on a
dedicated system is a good example of this), this
option (rpc_zmq_proxy_inprocess) allows that
operator to force ZmqProxy to run inside the
service manager process as a greenthread.

When proxy_inprocess is used, IPC sockets
are no longer created and ZeroMQ internally
delivers messages through the inproc transport.

This change also removes the setup/teardown of
the ZmqProxy from unit tests by simply enablig
this new option.

Change-Id: Ide05c06f3f6120ebc88a62cacc0bedd49305763e