Include host on IPC topics
Topics were being stripped of the host
value for IPC messaging. This caused
'base' topics to be round-robined or
fanned-out to all services running on
the system consuming messages for that
base topic, regardless of their host parameter.

MatchMakerLocalhost and LocalhostExchange
now take an optional host parameter.

_get_matchmaker now passes arguments to
the selected matchmaker module (generally
for unit testing purposes)

Change-Id: I939c24397e58492fc16561666aed3ca891325e9c