Faster ZmqProxy PUB socket creation
Because we could not verify if there were
subscribers to a socket and socket creation /
subscription was not immediate, a sleep(.5) was
used before sending messages on newly created
PUB sockets.

In production this is minor as it would only
cause a delay on the first received fanout and
message reply upon start of the -rpc-zmq-recever
process.

In unit testing, this caused an additional 500ms
delay on the running of each unit test for impl_zmq
code.

This patch switches from PUB sockets to the newer
XPUB socket type which allows us to wait for a subscriber
before sending messages.

This patch pushes the ZeroMQ requirement from 2.2+ to
ZeroMQ 3.2+

Change-Id: Ib94cb7894798882ce03a449c3d91b6d9bf5389df